Implemented closures in VM
Side note: I was actually in disbelief when this worked. I was debugging a 'Malformed bytecode' error which was showing the parser was incorrect, still didn't believe the VM code was correct... and then lo and behold I fixed one thing in the parser - basically at a guess - and it worked... Tests that had not passed since I introduced functions, all started passing. One last test was missing a `\n` on the expected output, but everything - every one of 373 tests, now passes. I am also now going to bed...