Fix various issues present with recursive data structures. Implement a recursive guard to `repr` and `str` implementations, which like Python, will prevent a recursive data structure from a stack overflow, and return a useful string representation.
Also prevents the creation of a recursive hash collection - a dictionary with itself as part of a key, and a set containing itself. The reason being `Hash` must take a borrow, as well as `insert` (or variants thereof). So we have to prevent it from existing in the first place.