Implement `complex` type, along with complex literals, and operators. A few major refactors that got included in the mix:
Replace and simplify the shared constant data structures to use a `Vec<Value>` instead of individual string, int, struct, function arrays. This has minor performance improvements when large literals are instantiated (due to a copy no longer being performed). It also simplifies the data structures passed between the VM <-> compiler during eval or incremental compiles.

Rewrite how disassembly is computed. For parser tests, `Opcode::disassemble` now is used for all tests, reducing boilerplate. In addition, the textual representation of a number of opcodes is simplified and made explicit from the Rust enum representation. This technically makes the disassembly less explicit for test correctness purposes, but it does make it more expressive as an assembly language, which is better for debugging.